+++ CHANGES.md
@@ -1,3 +1,7 @@
+# unreleased
+
+* better error message for `TileOutsideBounds` errors (author @abarciauskas-bgse, https://github.com/cogeotiff/rio-tiler/pull/712)
+
 # 6.6.1 (2024-05-17)
 
 * fix/support `scale/offset` indexes selection (author @jddeal, https://github.com/cogeotiff/rio-tiler/pull/709)