+++ CHANGES.md
@@ -1,6 +1,7 @@
 # unreleased
 
 * better error message for `TileOutsideBounds` errors (author @abarciauskas-bgse, https://github.com/cogeotiff/rio-tiler/pull/712)
+* handle of inverted latitude in `reader.point` (author @georgespill, https://github.com/cogeotiff/rio-tiler/pull/716)
 
 # 6.6.1 (2024-05-17)